feat: lombard wl asset and spot

+++ scripts/deploy-lrt-market.s.sol
@@ -88,7 +88,7 @@ contract DeployMarket is Utils {
     for (uint i = 0; i < config.feedSigners.length; ++i) {
       market.spotFeed.addSigner(config.feedSigners[i], true);
     }
-    market.spotFeed.setRequiredSigners(config.feedSigners.length);
+    market.spotFeed.setRequiredSigners(uint8(config.feedSigners.length));
 
     market.base = new WLWrappedERC20Asset(deployment.subAccounts, IERC20Metadata(marketERC20));
   }
